pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finnews_sentiment"
version = "0.1.0"
description = "Finance news ETL + sentiment analysis + backtest"
requires-python = ">=3.10"
dependencies = [
  "pydantic>=2.6",
  "python-dotenv>=1.0",
  "pyyaml>=6.0",
  "requests>=2.32",
  "beautifulsoup4>=4.12",
  "feedparser>=6.0",
  "newspaper3k>=0.2.8",
  "transformers>=4.44",
  "torch>=2.2",
  "nltk>=3.9",
  "vaderSentiment>=3.3.2",
  "spacy>=3.7",
  "yfinance>=0.2",
  "pandas>=2.2",
  "numpy>=1.26",
  "SQLAlchemy>=2.0",
  "matplotlib>=3.8",
  "scikit-learn>=1.5",
  "fastapi>=0.112",
  "uvicorn[standard]>=0.30",
  "streamlit>=1.36",
]

[tool.setuptools]
packages = ["finnews_sentiment"]
